+++ lib/src/isolate.dart
@@ -290,6 +290,10 @@ class VMIsolateRef {
         "resume", step == VMStep.resume ? {} : {"step": step._value});
   }
 
+  /// Sets the pause behaviour for exceptions.
+  Future setExceptionPauseMode(VMExceptionPauseMode mode) =>
+      _scope.sendRequest("setExceptionPauseMode", {"mode": mode._value});
+
   /// Sets the [name] of the isolate.
   ///
   /// Note that since this object is immutable, it needs to be reloaded to see
@@ -399,6 +403,9 @@ class VMIsolateRef {
 
 /// A full isolate on the remote VM.
 class VMIsolate extends VMIsolateRef {
+  /// The current pause on exception mode for this isolate.
+  final VMExceptionPauseMode exceptionPauseMode;
+
   /// The time that the isolate started running.
   final DateTime startTime;
 
@@ -433,7 +440,9 @@ class VMIsolate extends VMIsolateRef {
   final List<String> extensionRpcs;
 
   VMIsolate._(Scope scope, Map json)
-      : startTime = new DateTime.fromMillisecondsSinceEpoch(
+      : exceptionPauseMode =
+            new VMExceptionPauseMode._(json["exceptionPauseMode"]),
+        startTime = new DateTime.fromMillisecondsSinceEpoch(
             // Prior to v3.0, this was emitted as a double rather than an int.
             json["startTime"].round()),
         livePorts = json["livePorts"],
@@ -506,6 +515,25 @@ class VMStep {
   final String _value;
 
   const VMStep._(this._value);
+
+  String toString() => _value;
+}
+
+/// An enum of exception pause behaviour for use in [VMIsolateRef.setExceptionPauseMode].
+class VMExceptionPauseMode {
+  /// The isolate will not pause on any exceptions.
+  static const none = const VMExceptionPauseMode._("None");
+
+  /// The isolate will pause on any unhandled exceptions.
+  static const unhandled = const VMExceptionPauseMode._("Unhandled");
+
+  /// The isolate will pause on all exceptions.
+  static const all = const VMExceptionPauseMode._("All");
+
+  /// The string name of the exception pause mode.
+  final String _value;
+
+  const VMExceptionPauseMode._(this._value);
 
   String toString() => _value;
 }